refactor Rest layer (#17063)
* refactor Rest layer

* remove dead code

* more minor cleaning
