+++ conans/client/rest/rest_client.py
@@ -3,21 +3,7 @@ from conans.client.rest.rest_client_v2 import RestV2Methods
 from conans.errors import AuthenticationException, ConanException
 
 
-class RestApiClientFactory(object):
-
-    def __init__(self, requester, config):
-        self._requester = requester
-        self._config = config
-        self._cached_capabilities = {}
-
-    def new(self, remote, token, refresh_token, custom_headers):
-        tmp = RestApiClient(remote, token, refresh_token, custom_headers,
-                            self._requester, self._config,
-                            self._cached_capabilities)
-        return tmp
-
-
-class RestApiClient(object):
+class RestApiClient:
     """
         Rest Api Client for handle remote.
     """
@@ -113,9 +99,6 @@ class RestApiClient(object):
 
     def remove_packages(self, prefs):
         return self._get_api().remove_packages(prefs)
-
-    def server_capabilities(self):
-        return self._get_api().server_capabilities()
 
     def get_recipe_revisions_references(self, ref):
         return self._get_api().get_recipe_revisions_references(ref)
